requirements.txt
langgraph
langgraph-api
langgraph-cli
langgraph-cli[inmem]
langchain_community
langchain
langchain-openai
langchain-core
langchain-google-genai
langchain-mcp-adapters
python-dotenv
ipykernel
pydantic
starlette
uvicorn
openai
google-ai-generativelanguage
fastmcp
firecrawl-py
sympy